refactor large files into smaller ones; add more tests
