+++ table/demo/demo.go
@@ -417,16 +417,15 @@ func main() {
 	// Nope. I want a Markdown Table now.
 	//==========================================================================
 	for _, line := range strings.Split(t.RenderMarkdown(), "\n") {
-		fmt.Printf("MD| %s\n", line)
+		fmt.Printf("Markdown | %s\n", line)
 	}
 	fmt.Println()
-	//MD| | # | First Name | Last Name | Salary |  |
-	//MD| | ---:| --- | --- | ---:| --- |
-	//MD| | 1 | Arya | Stark | 3000 |  |
-	//MD| | 20 | Jon | Snow | 2000 | You know nothing, Jon Snow! |
-	//MD| | 300 | Tyrion | Lannister | 5000 |  |
-	//MD| |  |  | Total | 10000 |  |
-	//MD|
+	//Markdown | | # | First Name | Last Name | Salary |  |
+	//Markdown | | ---:| --- | --- | ---:| --- |
+	//Markdown | | 1 | Arya | Stark | 3000 |  |
+	//Markdown | | 20 | Jon | Snow | 2000 | You know nothing, Jon Snow! |
+	//Markdown | | 300 | Tyrion | Lannister | 5000 |  |
+	//Markdown | |  |  | Total | 10000 |  |
 	//==========================================================================
 
 	//==========================================================================
