+++ table/render_test.go
@@ -1,0 +1,65 @@
+package table
+
+import (
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestTable_Render(t *testing.T) {
+	tw := NewWriter()
+	tw.AppendHeader(testHeader)
+	tw.AppendRows(testRows)
+	tw.AppendRow(testRowMultiLine)
+	tw.AppendFooter(testFooter)
+	tw.SetAlign(testAlign)
+	tw.SetCaption(testCaption)
+	tw.SetStyle(styleTest)
+
+	expectedOut := `(-----^------------^-----------^--------^-----------------------------)
+[<  #>|<FIRST NAME>|<LAST NAME>|<SALARY>|<                           >]
+{-----+------------+-----------+--------+-----------------------------}
+[<  1>|<Arya      >|<Stark    >|<  3000>|<                           >]
+[< 20>|<Jon       >|<Snow     >|<  2000>|<You know nothing, Jon Snow!>]
+[<300>|<Tyrion    >|<Lannister>|<  5000>|<                           >]
+[<  0>|<Winter    >|<Is       >|<     0>|<Coming.                    >]
+[<   >|<          >|<         >|<      >|<The North Remembers!       >]
+{-----+------------+-----------+--------+-----------------------------}
+[<   >|<          >|<TOTAL    >|< 10000>|<                           >]
+\-----v------------v-----------v--------v-----------------------------/
+test-caption`
+
+	assert.Equal(t, expectedOut, tw.Render())
+}
+
+func TestTable_RenderColored(t *testing.T) {
+	tw := NewWriter()
+	tw.AppendHeader(testHeader)
+	tw.AppendRows(testRows)
+	tw.AppendRow(testRowMultiLine)
+	tw.AppendFooter(testFooter)
+	tw.SetAlign(testAlign)
+	tw.SetCaption(testCaption)
+	tw.SetColors(testColors)
+	tw.SetColorsFooter(testColorsFooter)
+	tw.SetColorsHeader(testColorsHeader)
+	tw.SetStyle(StyleRounded)
+
+	expectedOut := []string{
+		"╭─────┬────────────┬───────────┬────────┬─────────────────────────────╮",
+		"│\x1b[91;1m   # \x1b[0m│\x1b[91;1m FIRST NAME \x1b[0m│\x1b[91;1m LAST NAME \x1b[0m│\x1b[91;1m SALARY \x1b[0m│                             │",
+		"├─────┼────────────┼───────────┼────────┼─────────────────────────────┤",
+		"│\x1b[32m   1 \x1b[0m│\x1b[32m Arya       \x1b[0m│\x1b[32m Stark     \x1b[0m│\x1b[32m   3000 \x1b[0m│\x1b[36m                             \x1b[0m│",
+		"│\x1b[32m  20 \x1b[0m│\x1b[32m Jon        \x1b[0m│\x1b[32m Snow      \x1b[0m│\x1b[32m   2000 \x1b[0m│\x1b[36m You know nothing, Jon Snow! \x1b[0m│",
+		"│\x1b[32m 300 \x1b[0m│\x1b[32m Tyrion     \x1b[0m│\x1b[32m Lannister \x1b[0m│\x1b[32m   5000 \x1b[0m│\x1b[36m                             \x1b[0m│",
+		"│\x1b[32m   0 \x1b[0m│\x1b[32m Winter     \x1b[0m│\x1b[32m Is        \x1b[0m│\x1b[32m      0 \x1b[0m│\x1b[36m Coming.                     \x1b[0m│",
+		"│\x1b[32m     \x1b[0m│\x1b[32m            \x1b[0m│\x1b[32m           \x1b[0m│\x1b[32m        \x1b[0m│\x1b[36m The North Remembers!        \x1b[0m│",
+		"├─────┼────────────┼───────────┼────────┼─────────────────────────────┤",
+		"│     │            │\x1b[94;1m TOTAL     \x1b[0m│\x1b[94;1m  10000 \x1b[0m│                             │",
+		"╰─────┴────────────┴───────────┴────────┴─────────────────────────────╯",
+		"test-caption",
+	}
+
+	assert.Equal(t, strings.Join(expectedOut, "\n"), tw.Render())
+}
