+++ table/table_test.go
@@ -10,18 +10,20 @@ import (
 )
 
 var (
-	testAlign        = []text.Align{text.AlignDefault, text.AlignLeft, text.AlignLeft, text.AlignRight}
-	testCaption      = "test-caption"
-	testCSSClass     = "test-css-class"
-	testColor        = text.Colors{color.FgGreen}
-	testColorHeader  = text.Colors{color.FgHiRed, color.Bold}
-	testColorFooter  = text.Colors{color.FgHiBlue, color.Bold}
-	testColors       = []text.Colors{testColor, testColor, testColor, testColor, {color.FgCyan}}
-	testColorsFooter = []text.Colors{{}, {}, testColorFooter, testColorFooter}
-	testColorsHeader = []text.Colors{testColorHeader, testColorHeader, testColorHeader, testColorHeader}
-	testFooter       = Row{"", "", "Total", 10000}
-	testHeader       = Row{"#", "First Name", "Last Name", "Salary"}
-	testRows         = []Row{
+	testAlign           = []text.Align{text.AlignDefault, text.AlignLeft, text.AlignLeft, text.AlignRight}
+	testCaption         = "test-caption"
+	testColor           = text.Colors{color.FgGreen}
+	testColorBoW        = text.Colors{color.FgBlack, color.BgWhite}
+	testColorHiRedBold  = text.Colors{color.FgHiRed, color.Bold}
+	testColorHiBlueBold = text.Colors{color.FgHiBlue, color.Bold}
+	testColorWoB        = text.Colors{color.FgWhite, color.BgBlack}
+	testColors          = []text.Colors{testColor, testColor, testColor, testColor, {color.FgCyan}}
+	testColorsFooter    = []text.Colors{{}, {}, testColorHiBlueBold, testColorHiBlueBold}
+	testColorsHeader    = []text.Colors{testColorHiRedBold, testColorHiRedBold, testColorHiRedBold, testColorHiRedBold}
+	testCSSClass        = "test-css-class"
+	testFooter          = Row{"", "", "Total", 10000}
+	testHeader          = Row{"#", "First Name", "Last Name", "Salary"}
+	testRows            = []Row{
 		{1, "Arya", "Stark", 3000},
 		{20, "Jon", "Snow", 2000, "You know nothing, Jon Snow!"},
 		{300, "Tyrion", "Lannister", 5000},
@@ -30,8 +32,6 @@ var (
 	testRowNewLines  = Row{0, "Valar", "Morghulis", 0, "Faceless\nMen"}
 	testRowPipes     = Row{0, "Valar", "Morghulis", 0, "Faceless|Men"}
 	testRowTabs      = Row{0, "Valar", "Morghulis", 0, "Faceless\tMen"}
-	testColors1      = text.Colors{color.FgWhite, color.BgBlack}
-	testColors2      = text.Colors{color.FgBlack, color.BgWhite}
 )
 
 type myMockOutputMirror struct {
@@ -112,168 +112,6 @@ func TestTable_AppendRows(t *testing.T) {
 	assert.Equal(t, 0, len(table.rowsHeader))
 }
 
-func TestTable_Render(t *testing.T) {
-	tw := NewWriter()
-	tw.AppendHeader(testHeader)
-	tw.AppendRows(testRows)
-	tw.AppendRow(testRowMultiLine)
-	tw.AppendFooter(testFooter)
-	tw.SetAlign(testAlign)
-	tw.SetCaption(testCaption)
-	tw.SetStyle(styleTest)
-
-	expectedOut := `(-----^------------^-----------^--------^-----------------------------)
-[<  #>|<FIRST NAME>|<LAST NAME>|<SALARY>|<                           >]
-{-----+------------+-----------+--------+-----------------------------}
-[<  1>|<Arya      >|<Stark    >|<  3000>|<                           >]
-[< 20>|<Jon       >|<Snow     >|<  2000>|<You know nothing, Jon Snow!>]
-[<300>|<Tyrion    >|<Lannister>|<  5000>|<                           >]
-[<  0>|<Winter    >|<Is       >|<     0>|<Coming.                    >]
-[<   >|<          >|<         >|<      >|<The North Remembers!       >]
-{-----+------------+-----------+--------+-----------------------------}
-[<   >|<          >|<TOTAL    >|< 10000>|<                           >]
-\-----v------------v-----------v--------v-----------------------------/
-test-caption`
-
-	assert.Equal(t, expectedOut, tw.Render())
-}
-
-func TestTable_RenderColored(t *testing.T) {
-	tw := NewWriter()
-	tw.AppendHeader(testHeader)
-	tw.AppendRows(testRows)
-	tw.AppendRow(testRowMultiLine)
-	tw.AppendFooter(testFooter)
-	tw.SetAlign(testAlign)
-	tw.SetCaption(testCaption)
-	tw.SetColors(testColors)
-	tw.SetColorsFooter(testColorsFooter)
-	tw.SetColorsHeader(testColorsHeader)
-	tw.SetStyle(StyleRounded)
-
-	expectedOut := []string{
-		"╭─────┬────────────┬───────────┬────────┬─────────────────────────────╮",
-		"│\x1b[91;1m   # \x1b[0m│\x1b[91;1m FIRST NAME \x1b[0m│\x1b[91;1m LAST NAME \x1b[0m│\x1b[91;1m SALARY \x1b[0m│                             │",
-		"├─────┼────────────┼───────────┼────────┼─────────────────────────────┤",
-		"│\x1b[32m   1 \x1b[0m│\x1b[32m Arya       \x1b[0m│\x1b[32m Stark     \x1b[0m│\x1b[32m   3000 \x1b[0m│\x1b[36m                             \x1b[0m│",
-		"│\x1b[32m  20 \x1b[0m│\x1b[32m Jon        \x1b[0m│\x1b[32m Snow      \x1b[0m│\x1b[32m   2000 \x1b[0m│\x1b[36m You know nothing, Jon Snow! \x1b[0m│",
-		"│\x1b[32m 300 \x1b[0m│\x1b[32m Tyrion     \x1b[0m│\x1b[32m Lannister \x1b[0m│\x1b[32m   5000 \x1b[0m│\x1b[36m                             \x1b[0m│",
-		"│\x1b[32m   0 \x1b[0m│\x1b[32m Winter     \x1b[0m│\x1b[32m Is        \x1b[0m│\x1b[32m      0 \x1b[0m│\x1b[36m Coming.                     \x1b[0m│",
-		"│\x1b[32m     \x1b[0m│\x1b[32m            \x1b[0m│\x1b[32m           \x1b[0m│\x1b[32m        \x1b[0m│\x1b[36m The North Remembers!        \x1b[0m│",
-		"├─────┼────────────┼───────────┼────────┼─────────────────────────────┤",
-		"│     │            │\x1b[94;1m TOTAL     \x1b[0m│\x1b[94;1m  10000 \x1b[0m│                             │",
-		"╰─────┴────────────┴───────────┴────────┴─────────────────────────────╯",
-		"test-caption",
-	}
-
-	assert.Equal(t, strings.Join(expectedOut, "\n"), tw.Render())
-}
-
-func TestTable_RenderCSV(t *testing.T) {
-	tw := NewWriter()
-	tw.AppendHeader(testHeader)
-	tw.AppendRows(testRows)
-	tw.AppendRow(testRowMultiLine)
-	tw.AppendRow(testRowTabs)
-	tw.AppendFooter(testFooter)
-
-	expectedOut := `#,First Name,Last Name,Salary,
-1,Arya,Stark,3000,
-20,Jon,Snow,2000,"You know nothing\, Jon Snow!"
-300,Tyrion,Lannister,5000,
-0,Winter,Is,0,"Coming.
-The North Remembers!"
-0,Valar,Morghulis,0,Faceless    Men
-,,Total,10000,`
-
-	assert.Equal(t, expectedOut, tw.RenderCSV())
-}
-
-func TestTable_RenderHTML(t *testing.T) {
-	tw := NewWriter()
-	tw.AppendHeader(testHeader)
-	tw.AppendRows(testRows)
-	tw.AppendRow(testRowMultiLine)
-	tw.AppendFooter(testFooter)
-	tw.SetVAlign([]text.VAlign{
-		text.VAlignDefault, text.VAlignDefault, text.VAlignDefault, text.VAlignBottom, text.VAlignBottom})
-
-	expectedOut := `<table class="go-pretty-table">
-  <thead>
-  <tr>
-    <th align="right">#</th>
-    <th>First Name</th>
-    <th>Last Name</th>
-    <th align="right" valign="bottom">Salary</th>
-    <th valign="bottom">&nbsp;</th>
-  </tr>
-  </thead>
-  <tbody>
-  <tr>
-    <td align="right">1</td>
-    <td>Arya</td>
-    <td>Stark</td>
-    <td align="right" valign="bottom">3000</td>
-    <td valign="bottom">&nbsp;</td>
-  </tr>
-  <tr>
-    <td align="right">20</td>
-    <td>Jon</td>
-    <td>Snow</td>
-    <td align="right" valign="bottom">2000</td>
-    <td valign="bottom">You know nothing, Jon Snow!</td>
-  </tr>
-  <tr>
-    <td align="right">300</td>
-    <td>Tyrion</td>
-    <td>Lannister</td>
-    <td align="right" valign="bottom">5000</td>
-    <td valign="bottom">&nbsp;</td>
-  </tr>
-  <tr>
-    <td align="right">0</td>
-    <td>Winter</td>
-    <td>Is</td>
-    <td align="right" valign="bottom">0</td>
-    <td valign="bottom">Coming.<br/>The North Remembers!</td>
-  </tr>
-  </tbody>
-  <tfoot>
-  <tr>
-    <td align="right">&nbsp;</td>
-    <td>&nbsp;</td>
-    <td>Total</td>
-    <td align="right" valign="bottom">10000</td>
-    <td valign="bottom">&nbsp;</td>
-  </tr>
-  </tfoot>
-</table>`
-
-	assert.Equal(t, expectedOut, tw.RenderHTML())
-}
-
-func TestTable_RenderMarkdown(t *testing.T) {
-	tw := NewWriter()
-	tw.AppendHeader(testHeader)
-	tw.AppendRows(testRows)
-	tw.AppendRow(testRowNewLines)
-	tw.AppendRow(testRowPipes)
-	tw.AppendFooter(testFooter)
-	tw.SetCaption(testCaption)
-
-	expectedOut := `| # | First Name | Last Name | Salary |  |
-| ---:| --- | --- | ---:| --- |
-| 1 | Arya | Stark | 3000 |  |
-| 20 | Jon | Snow | 2000 | You know nothing, Jon Snow! |
-| 300 | Tyrion | Lannister | 5000 |  |
-| 0 | Valar | Morghulis | 0 | Faceless<br>Men |
-| 0 | Valar | Morghulis | 0 | Faceless\|Men |
-|  |  | Total | 10000 |  |
-_test-caption_`
-
-	assert.Equal(t, expectedOut, tw.RenderMarkdown())
-}
-
 func TestTable_SetAlign(t *testing.T) {
 	table := Table{}
 	assert.Nil(t, table.align)
@@ -311,7 +149,7 @@ func TestTable_SetColors(t *testing.T) {
 	assert.Empty(t, table.colorsFooter)
 	assert.Empty(t, table.colorsHeader)
 
-	table.SetColors([]text.Colors{testColors1, testColors2})
+	table.SetColors([]text.Colors{testColorWoB, testColorBoW})
 	assert.NotEmpty(t, table.colors)
 	assert.Empty(t, table.colorsFooter)
 	assert.Empty(t, table.colorsHeader)
@@ -324,7 +162,7 @@ func TestTable_SetColorsFooter(t *testing.T) {
 	assert.Empty(t, table.colorsFooter)
 	assert.Empty(t, table.colorsHeader)
 
-	table.SetColorsFooter([]text.Colors{testColors1, testColors2})
+	table.SetColorsFooter([]text.Colors{testColorWoB, testColorBoW})
 	assert.Empty(t, table.colors)
 	assert.NotEmpty(t, table.colorsFooter)
 	assert.Empty(t, table.colorsHeader)
@@ -337,7 +175,7 @@ func TestTable_SetColorsHeader(t *testing.T) {
 	assert.Empty(t, table.colorsFooter)
 	assert.Empty(t, table.colorsHeader)
 
-	table.SetColorsHeader([]text.Colors{testColors1, testColors2})
+	table.SetColorsHeader([]text.Colors{testColorWoB, testColorBoW})
 	assert.Empty(t, table.colors)
 	assert.Empty(t, table.colorsFooter)
 	assert.NotEmpty(t, table.colorsHeader)
